Splitting Map component away from App starting point
